implemented mark controller crud.

+++ src/main/java/com/dk/repository/MarkRepository.java
@@ -3,5 +3,5 @@ package com.dk.repository;
 import com.dk.model.Mark;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MarkRepository extends CrudRepository<Mark, Integer> {
+public interface MarkRepository extends CrudRepository<Mark, Long> {
 }
