+++ src/main/java/com/dk/repository/ReportRepository.java
@@ -3,5 +3,5 @@ package com.dk.repository;
 import com.dk.model.Report;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReportRepository extends CrudRepository<Report, Integer> {
+public interface ReportRepository extends CrudRepository<Report, Long> {
 }
